@@ -15,17 +15,12 @@ import (
 	"time"
 
 	"github.com/johncferguson/gotunnel/internal/cert"
-	"github.com/johncferguson/gotunnel/internal/dnsserver"
 	"github.com/johncferguson/gotunnel/internal/tunnel"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestMain(m *testing.M) {
-	// Create a context with timeout for the entire test suite
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
 	// Setup test environment
 	tempDir, err := os.MkdirTemp("", "gotunnel-test-*")
 	if err != nil {
@@ -33,15 +28,6 @@ func TestMain(m *testing.M) {
 		os.Exit(1)
 	}
 	defer os.RemoveAll(tempDir)
-
-	// Run cleanup after tests finish
-	defer func() {
-		shutdownCtx, shutdownCancel := context.WithTimeout(ctx, 2*time.Second)
-		defer shutdownCancel()
-		if err := dnsserver.Shutdown(); err != nil {
-			log.Printf("Error shutting down DNS server: %v", err)
-		}
-	}()
 
 	os.Exit(m.Run())
 }
@@ -52,7 +38,6 @@ func setupTestServer(t *testing.T) (*http.Server, int) {
 	// Create a test HTTP server with a random available port
 	listener, err := net.Listen("tcp", ":0")
 	require.NoError(t, err)
-
 	port := listener.Addr().(*net.TCPAddr).Port
 
 	srv := &http.Server{
@@ -67,108 +52,132 @@ func setupTestServer(t *testing.T) (*http.Server, int) {
 		}
 	}()
 
-	// Wait for server to start
-	time.Sleep(100 * time.Millisecond)
 	return srv, port
 }
 
 func setupTestServerWithCleanup(t *testing.T) (*http.Server, int, func()) {
 	srv, port := setupTestServer(t)
 	return srv, port, func() {
-		ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
+		shutdownCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 		defer cancel()
-		if err := srv.Shutdown(ctx); err != nil {
+		if err := srv.Shutdown(shutdownCtx); err != nil {
 			t.Logf("Error shutting down test server: %v", err)
 		}
 	}
 }
 
 func setupTunnelManagerWithCleanup(t *testing.T) (*tunnel.Manager, func()) {
-	certManager := cert.New(os.TempDir())
+	// Create a temp directory for certs and hosts backup
+	tmpDir, err := os.MkdirTemp("", "gotunnel-test-*")
+	require.NoError(t, err)
+
+	// Create cert manager with temp dir for certs
+	certManager := cert.New(tmpDir)
+
+	// Create tunnel manager with temp file for hosts backup
 	manager := tunnel.NewManager(certManager)
+	manager.SetHostsBackupDir(filepath.Join(tmpDir, "hosts.bak"))
+
 	return manager, func() {
-		ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-		defer cancel()
-		if err := manager.Close(ctx); err != nil {
-			t.Logf("Error closing tunnel manager: %v", err)
-		}
-		// Ensure DNS server is cleaned up
-		if err := dnsserver.Shutdown(); err != nil {
-			t.Logf("Error shutting down DNS server: %v", err)
-		}
+		// Cleanup
+		os.RemoveAll(tmpDir)
 	}
 }
 
 func TestTunnelCreation(t *testing.T) {
-	// Create test server with random port
-	_, port, cleanupSrv := setupTestServerWithCleanup(t)
-	defer cleanupSrv()
+	// Remove the privilege check
+	// if os.Getuid() != 0 {
+	//     t.Skip("Skipping test - requires root privileges")
+	// }
 
-	// Create tunnel manager
-	manager, cleanupManager := setupTunnelManagerWithCleanup(t)
-	defer cleanupManager()
+	// Create tunnel manager first
+	manager, cleanup := setupTunnelManagerWithCleanup(t)
+	defer cleanup()
 
 	tests := []struct {
-		name      string
-		domain    string
-		port      int
-		httpsPort int
-		https     bool
-		wantErr   bool
+		name    string
+		domain  string
+		https   bool
+		wantErr bool
 	}{
 		{
-			name:      "Basic HTTP Tunnel",
-			domain:    "test-http",
-			port:      port,
-			httpsPort: port + 1,
-			https:     false,
-			wantErr:   false,
+			name:    "Basic HTTP Tunnel",
+			domain:  "test-http",
+			https:   false,
+			wantErr: false,
 		},
 		{
-			name:      "HTTPS Tunnel",
-			domain:    "test-https",
-			port:      port + 2,
-			httpsPort: port + 3,
-			https:     true,
-			wantErr:   false,
+			name:    "HTTPS Tunnel",
+			domain:  "test-https",
+			https:   true,
+			wantErr: false,
 		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			// Skip HTTPS tests if mkcert is not available or we don't have permissions
+			// Skip HTTPS tests if mkcert is not available
 			if tt.https {
-				if _, err := exec.LookPath("mkcert"); err != nil || os.Getuid() != 0 {
-					t.Skip("Skipping HTTPS test - mkcert not available or not running as root")
+				if _, err := exec.LookPath("mkcert"); err != nil {
+					t.Skip("Skipping HTTPS test - mkcert not available")
 				}
 			}
 
-			ctx := context.Background()
+			// Create test server for this test case (this is our target)
+			srv, targetPort := setupTestServer(t)
+			defer func() {
+				ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+				defer cancel()
+				if err := srv.Shutdown(ctx); err != nil {
+					t.Logf("Error shutting down test server: %v", err)
+				}
+			}()
 
-			err := manager.StartTunnel(ctx, tt.port, tt.domain, tt.https, tt.httpsPort)
+			// Get available port for tunnel to listen on
+			tunnelListener, err := net.Listen("tcp", ":0")
+			require.NoError(t, err)
+			tunnelPort := tunnelListener.Addr().(*net.TCPAddr).Port
+			tunnelListener.Close() // Close immediately to free the port
+
+			var httpsPort int
+			if tt.https {
+				httpsListener, err := net.Listen("tcp", ":0")
+				require.NoError(t, err)
+				httpsPort = httpsListener.Addr().(*net.TCPAddr).Port
+				httpsListener.Close() // Close immediately to free the port
+			}
+
+			// Start tunnel with custom ports for testing
+			// targetPort = backend app port, tunnelPort/httpsPort = tunnel listen ports
+			if tt.https {
+				err = manager.StartTunnelWithPorts(context.Background(), targetPort, tt.domain, true, 0, httpsPort)
+			} else {
+				err = manager.StartTunnelWithPorts(context.Background(), targetPort, tt.domain, false, tunnelPort, 0)
+			}
 			if tt.wantErr {
 				assert.Error(t, err)
 				return
 			}
+			require.NoError(t, err)
 
-			assert.NoError(t, err)
+			// Ensure tunnel is cleaned up
 			defer func() {
-				ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
+				ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 				defer cancel()
 				if err := manager.StopTunnel(ctx, tt.domain+".local"); err != nil {
 					t.Logf("Error stopping tunnel: %v", err)
 				}
 			}()
 
-			// Give DNS time to propagate (reduced from 500ms)
-			time.Sleep(100 * time.Millisecond)
+			// Give DNS time to propagate
+			time.Sleep(2 * time.Second)
 
-			// Test the tunnel...
+			// Test the tunnel by connecting to tunnelPort (HTTP) or httpsPort (HTTPS)
 			protocol := "http"
-			testPort := tt.port
+			testPort := tunnelPort
 			if tt.https {
 				protocol = "https"
-				testPort = tt.httpsPort
+				testPort = httpsPort
 			}
 
 			client := &http.Client{
@@ -177,12 +186,27 @@ func TestTunnelCreation(t *testing.T) {
 						InsecureSkipVerify: true,
 					},
 				},
-				Timeout: 2 * time.Second, // Add timeout to prevent hanging
+				Timeout: 5 * time.Second,
 			}
-			resp, err := client.Get(fmt.Sprintf("%s://%s.local:%d", protocol, tt.domain, testPort))
-			require.NoError(t, err)
-			defer resp.Body.Close()
 
+			// Connect directly to the tunnel listener for testing
+			// In production, this would go through domain resolution
+			var resp *http.Response
+			var lastErr error
+			for i := 0; i < 5; i++ {
+				// Test by connecting directly to the tunnel port
+				resp, err = client.Get(fmt.Sprintf("%s://127.0.0.1:%d", protocol, testPort))
+				if err == nil {
+					break
+				}
+				lastErr = err
+				time.Sleep(1 * time.Second)
+			}
+			if lastErr != nil {
+				t.Fatalf("Failed to connect after retries: %v", lastErr)
+			}
+
+			defer resp.Body.Close()
 			body, err := io.ReadAll(resp.Body)
 			require.NoError(t, err)
 			assert.Equal(t, "Hello from test server!", string(body))
@@ -191,28 +215,69 @@ func TestTunnelCreation(t *testing.T) {
 }
 
 func TestTunnelManagement(t *testing.T) {
+	// Skip if not running as root since we need to modify hosts file
+	if os.Getuid() != 0 {
+		t.Skip("Skipping test - requires root privileges")
+	}
+
 	// Create temp directory for test
 	tempDir, err := os.MkdirTemp("", "gotunnel-mgmt-test-*")
 	require.NoError(t, err)
 	defer os.RemoveAll(tempDir)
 
-	certManager := cert.New(filepath.Join(tempDir, "certs"))
+	// Create cert manager with temp dir for certs
+	certManager := cert.New(tempDir)
+
+	// Create tunnel manager with temp file for hosts backup
 	manager := tunnel.NewManager(certManager)
+	manager.SetHostsBackupDir(filepath.Join(tempDir, "hosts.bak"))
+
+	// Test tunnel management operations
 	ctx := context.Background()
 
-	// Test multiple tunnel creation
-	domains := []string{"test1.local", "test2.local", "test3.local"}
+	// Create multiple test servers and tunnels
+	var servers []*http.Server
+	var ports []int
+	for i := 0; i < 3; i++ {
+		srv, port := setupTestServer(t)
+		servers = append(servers, srv)
+		ports = append(ports, port)
+	}
+
+	// Cleanup test servers
+	defer func() {
+		for _, srv := range servers {
+			ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+			defer cancel()
+			if err := srv.Shutdown(ctx); err != nil {
+				t.Logf("Error shutting down test server: %v", err)
+			}
+		}
+	}()
+
+	// Start multiple tunnels
+	domains := []string{"test1", "test2", "test3"}
 	for i, domain := range domains {
-		err := manager.StartTunnel(ctx, 8080+i, domain, false, 9080+i)
+		err := manager.StartTunnel(ctx, ports[i], domain, false, 0)
 		require.NoError(t, err)
 	}
 
 	// Test tunnel listing
 	tunnels := manager.ListTunnels()
 	assert.Len(t, tunnels, len(domains))
+	for _, domain := range domains {
+		found := false
+		for _, tunnel := range tunnels {
+			if tunnel["domain"].(string) == domain+".local" {
+				found = true
+				break
+			}
+		}
+		assert.True(t, found, "Tunnel for domain %s not found", domain)
+	}
 
 	// Test individual tunnel stopping
-	err = manager.StopTunnel(ctx, domains[0])
+	err = manager.StopTunnel(ctx, domains[0]+".local")
 	require.NoError(t, err)
 	tunnels = manager.ListTunnels()
 	assert.Len(t, tunnels, len(domains)-1)
@@ -225,50 +290,60 @@ func TestTunnelManagement(t *testing.T) {
 }
 
 func TestErrorHandling(t *testing.T) {
+	// Skip if not running as root since we need to modify hosts file
+	if os.Getuid() != 0 {
+		t.Skip("Skipping test - requires root privileges")
+	}
+
 	tempDir, err := os.MkdirTemp("", "gotunnel-error-test-*")
 	require.NoError(t, err)
 	defer os.RemoveAll(tempDir)
 
-	certManager := cert.New(filepath.Join(tempDir, "certs"))
+	certManager := cert.New(tempDir)
 	manager := tunnel.NewManager(certManager)
-	ctx := context.Background()
+	manager.SetHostsBackupDir(filepath.Join(tempDir, "hosts.bak"))
 
 	tests := []struct {
 		name    string
-		fn      func() error
-		wantErr bool
+		setup   func(t *testing.T) error
+		wantErr string
 	}{
 		{
 			name: "Invalid Port",
-			fn: func() error {
-				return manager.StartTunnel(ctx, -1, "test", false, 9080)
+			setup: func(t *testing.T) error {
+				return manager.StartTunnel(context.Background(), -1, "test", false, 0)
 			},
-			wantErr: true,
+			wantErr: "invalid port",
 		},
 		{
 			name: "Empty Domain",
-			fn: func() error {
-				return manager.StartTunnel(ctx, 8080, "", false, 9080)
+			setup: func(t *testing.T) error {
+				return manager.StartTunnel(context.Background(), 8080, "", false, 0)
 			},
-			wantErr: true,
+			wantErr: "empty domain",
 		},
 		{
 			name: "Stop Non-existent Tunnel",
-			fn: func() error {
-				return manager.StopTunnel(ctx, "nonexistent")
+			setup: func(t *testing.T) error {
+				return manager.StopTunnel(context.Background(), "nonexistent.local")
 			},
-			wantErr: true,
+			wantErr: "tunnel not found",
+		},
+		{
+			name: "Invalid Hosts Backup Path",
+			setup: func(t *testing.T) error {
+				manager.SetHostsBackupDir("/nonexistent/dir/hosts.bak")
+				return manager.StartTunnel(context.Background(), 8080, "test", false, 0)
+			},
+			wantErr: "failed to backup hosts file",
 		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			err := tt.fn()
-			if tt.wantErr {
-				assert.Error(t, err)
-			} else {
-				assert.NoError(t, err)
-			}
+			err := tt.setup(t)
+			require.Error(t, err)
+			assert.Contains(t, err.Error(), tt.wantErr)
 		})
 	}
 }
