@@ -19,40 +19,53 @@ import (
 
 	"github.com/johncferguson/gotunnel/internal/cert"
 	"github.com/johncferguson/gotunnel/internal/dnsserver"
+	"github.com/johncferguson/gotunnel/internal/proxy"
 )
 
 const (
-	hostsFile = "/etc/hosts"
+	defaultHostsFile = "/etc/hosts"
 )
 
+// For testing purposes - allow overriding the hosts file path
+var hostsFile = defaultHostsFile
+
 type Tunnel struct {
-	Port      int
-	HTTPSPort int
-	Domain    string
-	TargetIP  string
-	HTTPS     bool
-	server    *http.Server
-	listener  net.Listener
-	done      chan struct{}
-	Cert      *tls.Certificate
+	Port        int    // Backend target port (where user's app runs)
+	HTTPPort    int    // Tunnel HTTP listen port (default 80)
+	HTTPSPort   int    // Tunnel HTTPS listen port (default 443) 
+	Domain      string
+	TargetIP    string
+	HTTPS       bool
+	server      *http.Server
+	listener    net.Listener
+	done        chan struct{}
+	Cert        *tls.Certificate
 }
 
 type Manager struct {
-	tunnels     map[string]*Tunnel
-	mu          sync.RWMutex
-	certManager *cert.CertManager
-	hostsBackup string
+	tunnels      map[string]*Tunnel
+	mu           sync.RWMutex
+	certManager  *cert.CertManager
+	hostsBackup  string
+	proxyManager *proxy.Manager
+	useProxy     bool
 }
 
 func NewManager(certManager *cert.CertManager) *Manager {
+	return NewManagerWithProxy(certManager, nil, false)
+}
+
+func NewManagerWithProxy(certManager *cert.CertManager, proxyManager *proxy.Manager, useProxy bool) *Manager {
 	// Initialize DNS server when creating a new manager
 	if err := dnsserver.StartDNSServer(); err != nil {
 		log.Printf("Warning: Failed to initialize DNS server: %v", err)
 	}
 
 	return &Manager{
-		tunnels:     make(map[string]*Tunnel),
-		certManager: certManager,
+		tunnels:      make(map[string]*Tunnel),
+		certManager:  certManager,
+		proxyManager: proxyManager,
+		useProxy:     useProxy,
 	}
 }
 
@@ -93,13 +106,59 @@ func (m *Manager) restoreHostsFile() error {
 	return nil
 }
 
-func (m *Manager) StartTunnel(ctx context.Context, port int, domain string, https bool, httpsPort int) error {
+// StartTunnelWithPorts starts a tunnel with custom listen ports (for testing)
+func (m *Manager) StartTunnelWithPorts(ctx context.Context, backendPort int, domain string, https bool, httpPort, httpsPort int) error {
+	// Set defaults if needed
+	if httpsPort == 0 {
+		httpsPort = 443
+	}
+	if httpPort == 0 {
+		httpPort = 80
+	}
+
+	return m.startTunnelInternal(ctx, backendPort, domain, https, httpPort, httpsPort)
+}
+
+// StartTunnel starts a tunnel with default ports (production use)
+func (m *Manager) StartTunnel(ctx context.Context, backendPort int, domain string, https bool, httpsPort int) error {
+	return m.StartTunnelWithPorts(ctx, backendPort, domain, https, 80, httpsPort)
+}
+
+func (m *Manager) startTunnelInternal(ctx context.Context, backendPort int, domain string, https bool, httpPort, httpsPort int) error {
 	m.mu.Lock()
 	defer m.mu.Unlock()
+
+	// Validate inputs
+	if backendPort <= 0 || backendPort > 65535 {
+		return fmt.Errorf("invalid backend port: %d", backendPort)
+	}
+	if domain == "" {
+		return fmt.Errorf("domain cannot be empty")
+	}
+	if httpPort <= 0 || httpPort > 65535 {
+		return fmt.Errorf("invalid HTTP port: %d", httpPort)
+	}
+	if httpsPort <= 0 || httpsPort > 65535 {
+		return fmt.Errorf("invalid HTTPS port: %d", httpsPort)
+	}
 
 	// Prevent duplicate tunnels for the same domain
 	if _, exists := m.tunnels[domain]; exists {
 		return fmt.Errorf("tunnel for domain %s already exists", domain)
+	}
+
+	// If using proxy, modify ports to avoid conflicts
+	tunnelHTTPPort := httpPort
+	tunnelHTTPSPort := httpsPort
+	
+	if m.useProxy && m.proxyManager != nil {
+		// Use high ports for actual tunnel, proxy will handle 80/443
+		// Start from 9080 to avoid conflicts with proxy on 8080
+		tunnelHTTPPort = 9080 + len(m.tunnels)  // Dynamic port allocation  
+		tunnelHTTPSPort = 9443 + len(m.tunnels)
+		
+		log.Printf("Using proxy mode: tunnel will run on ports %d/%d, accessible via proxy on %d/%d", 
+			tunnelHTTPPort, tunnelHTTPSPort, httpPort, httpsPort)
 	}
 
 	// Convert domain to .local if not already
@@ -109,8 +168,9 @@ func (m *Manager) StartTunnel(ctx context.Context, port int, domain string, http
 
 	// Create new tunnel instance
 	tunnel := &Tunnel{
-		Port:      port,
-		HTTPSPort: httpsPort,
+		Port:      backendPort,      // Backend target port (where user's app runs)
+		HTTPPort:  tunnelHTTPPort,   // Tunnel HTTP listen port (may be high port if using proxy)
+		HTTPSPort: tunnelHTTPSPort,  // Tunnel HTTPS listen port (may be high port if using proxy)
 		Domain:    domain,
 		TargetIP:  "127.0.0.1",
 		HTTPS:     https,
@@ -130,24 +190,27 @@ func (m *Manager) StartTunnel(ctx context.Context, port int, domain string, http
 		return fmt.Errorf("failed to start tunnel: %w", err)
 	}
 
-	// Register the domain with mDNS
-	servicePort := port
-	if https {
-		servicePort = httpsPort
-	}
-	if err := dnsserver.RegisterDomain(domain, servicePort); err != nil {
-		// Stop the tunnel we just started since mDNS registration failed
-		if stopErr := tunnel.stop(ctx); stopErr != nil {
-			return fmt.Errorf("failed to stop tunnel after mDNS registration failed: %w", stopErr)
-		}
-		return fmt.Errorf("failed to register domain with mDNS: %w", err)
-	}
-
 	// Add to internal map for tracking
 	m.tunnels[domain] = tunnel
 
-	// Create hosts file backup before first modification
-	if len(m.tunnels) == 1 {
+	// Register with proxy if using proxy mode
+	if m.useProxy && m.proxyManager != nil {
+		route := &proxy.Route{
+			Domain:     domain,
+			TargetHost: "127.0.0.1",
+			TargetPort: tunnel.HTTPPort, // Proxy routes to tunnel's actual port
+			HTTPS:      https,
+		}
+		
+		if err := m.proxyManager.AddRoute(route); err != nil {
+			log.Printf("Warning: Failed to register proxy route: %v", err)
+		} else {
+			log.Printf("✅ Registered proxy route: %s -> 127.0.0.1:%d", domain, tunnel.HTTPPort)
+		}
+	}
+
+	// Create hosts file backup before first modification (only if not using proxy)
+	if !m.useProxy && len(m.tunnels) == 1 {
 		if err := m.backupHostsFile(); err != nil {
 			return fmt.Errorf("failed to backup hosts file: %w", err)
 		}
@@ -198,6 +261,22 @@ func (m *Manager) StopTunnel(ctx context.Context, domain string) error {
 		return fmt.Errorf("failed to stop tunnel: %w", err)
 	}
 
+	// Remove from hosts file (only if not using proxy mode)
+	if !m.useProxy {
+		if err := removeFromHostsFile(domain); err != nil {
+			log.Printf("Warning: Failed to remove from hosts file: %v", err)
+		}
+	}
+
+	// Remove from proxy if using proxy mode
+	if m.useProxy && m.proxyManager != nil {
+		if err := m.proxyManager.RemoveRoute(domain); err != nil {
+			log.Printf("Warning: Failed to remove proxy route: %v", err)
+		} else {
+			log.Printf("🗑️  Removed proxy route: %s", domain)
+		}
+	}
+
 	// Unregister from mDNS
 	if err := dnsserver.UnregisterDomain(domain); err != nil {
 		return fmt.Errorf("failed to unregister domain from mDNS: %w", err)
@@ -210,23 +289,25 @@ func (m *Manager) StopTunnel(ctx context.Context, domain string) error {
 
 func (t *Tunnel) stop(ctx context.Context) error {
 	if t.server != nil {
+		// Server shutdown should gracefully close the listener
 		if err := t.server.Shutdown(ctx); err != nil {
+			// If graceful shutdown fails, force close the listener
+			if t.listener != nil {
+				t.listener.Close()
+			}
 			return fmt.Errorf("error shutting down server: %w", err)
 		}
-	}
-
-	if t.listener != nil {
-		// Force close the listener to unblock any pending accepts
+		t.server = nil
+	} else if t.listener != nil {
+		// Only close listener directly if server wasn't running
 		if err := t.listener.Close(); err != nil {
 			return fmt.Errorf("error closing listener: %w", err)
 		}
-		t.listener = nil
 	}
+	t.listener = nil
 
-	// Remove from hosts file
-	if err := removeFromHostsFile(t.Domain); err != nil {
-		return fmt.Errorf("failed to remove from hosts file: %w", err)
-	}
+	// Remove from hosts file (will be handled by manager for proxy mode)
+	// Note: This is called from manager which handles proxy mode appropriately
 
 	close(t.done)
 	return nil
@@ -280,9 +361,22 @@ func (m *Manager) startTunnel(t *Tunnel) error {
 	ip := dnsserver.GetOutboundIP()
 	t.TargetIP = ip.String()
 
-	// Update /etc/hosts file
-	if err := updateHostsFile(t.Domain); err != nil {
-		return fmt.Errorf("failed to update hosts file: %w", err)
+	// Update /etc/hosts file (skip if using proxy mode)
+	if !m.useProxy {
+		if err := updateHostsFile(t.Domain); err != nil {
+			return fmt.Errorf("failed to update hosts file: %w", err)
+		}
+	} else {
+		log.Printf("Skipping hosts file update (using proxy mode)")
+	}
+
+	// Register domain with DNS server (use tunnel listen port, not backend port)
+	listenPort := t.HTTPPort
+	if t.HTTPS {
+		listenPort = t.HTTPSPort
+	}
+	if err := dnsserver.RegisterDomain(t.Domain, listenPort); err != nil {
+		return fmt.Errorf("failed to register domain: %w", err)
 	}
 
 	// Create reverse proxy
@@ -313,8 +407,9 @@ func (m *Manager) startTunnel(t *Tunnel) error {
 	// Initialize done channel
 	t.done = make(chan struct{})
 
-	// Explicitly bind to all interfaces with the correct port
+	// Explicitly bind to all interfaces with the tunnel listen port
 	if t.HTTPS {
+		// Listen on HTTPS port for the tunnel (default 443)
 		baseListener, err = config.Listen(context.Background(), "tcp", fmt.Sprintf("0.0.0.0:%d", t.HTTPSPort))
 		if err != nil {
 			return fmt.Errorf("failed to create HTTPS listener: %w", err)
@@ -338,7 +433,8 @@ func (m *Manager) startTunnel(t *Tunnel) error {
 
 		t.listener = tls.NewListener(baseListener, tlsConfig)
 	} else {
-		baseListener, err = config.Listen(context.Background(), "tcp", fmt.Sprintf("0.0.0.0:%d", t.Port))
+		// Listen on HTTP port for the tunnel (default 80), not backend port
+		baseListener, err = config.Listen(context.Background(), "tcp", fmt.Sprintf("0.0.0.0:%d", t.HTTPPort))
 		if err != nil {
 			return fmt.Errorf("failed to create HTTP listener: %w", err)
 		}
@@ -398,9 +494,14 @@ func (m *Manager) Close(ctx context.Context) error {
 	return nil
 }
 
+func (m *Manager) SetHostsBackupDir(dir string) {
+	m.mu.Lock()
+	defer m.mu.Unlock()
+	m.hostsBackup = dir
+}
+
 // updateHostsFile adds or updates an entry in /etc/hosts
 func updateHostsFile(domain string) error {
-	hostsFile := "/etc/hosts"
 
 	// Read current hosts file
 	content, err := os.ReadFile(hostsFile)
@@ -429,7 +530,6 @@ func updateHostsFile(domain string) error {
 
 // removeFromHostsFile removes an entry from /etc/hosts
 func removeFromHostsFile(domain string) error {
-	hostsFile := "/etc/hosts"
 
 	// Read current hosts file
 	content, err := os.ReadFile(hostsFile)
