@@ -20,10 +20,25 @@ func setupTestManager(t *testing.T) (*Manager, string, func()) {
 
 	certManager := cert.New(filepath.Join(tempDir, "certs"))
 	manager := NewManager(certManager)
+	
+	// Set a temp directory for hosts backup for testing
+	hostsBackupFile := filepath.Join(tempDir, "hosts.backup")
+	manager.SetHostsBackupDir(hostsBackupFile)
+	
+	// Create a mock hosts file for testing
+	mockHostsFile := filepath.Join(tempDir, "mock_hosts")
+	err = os.WriteFile(mockHostsFile, []byte("127.0.0.1\tlocalhost\n"), 0644)
+	require.NoError(t, err)
+	
+	// Override the global hosts file variable for testing
+	originalHostsFile := hostsFile
+	hostsFile = mockHostsFile
 
 	cleanup := func() {
 		ctx := context.Background()
 		manager.Stop(ctx)
+		// Restore original hosts file path
+		hostsFile = originalHostsFile
 		os.RemoveAll(tempDir)
 	}
 
@@ -55,11 +70,12 @@ func TestStartAndStopTunnel(t *testing.T) {
 
 	ctx := context.Background()
 	domain := "test-tunnel.local"
-	port := 8080
+	backendPort := 8080
+	httpPort := 8180   // Use different port to avoid conflicts with other tests
 	httpsPort := 8443
 
-	// Start tunnel
-	err := manager.StartTunnel(ctx, port, domain, false, httpsPort)
+	// Start tunnel with custom ports
+	err := manager.StartTunnelWithPorts(ctx, backendPort, domain, false, httpPort, httpsPort)
 	require.NoError(t, err)
 
 	// Verify tunnel is created
@@ -67,8 +83,9 @@ func TestStartAndStopTunnel(t *testing.T) {
 	tunnel, exists := manager.tunnels[domain]
 	manager.mu.RUnlock()
 	assert.True(t, exists)
-	assert.Equal(t, port, tunnel.Port)
+	assert.Equal(t, backendPort, tunnel.Port)
 	assert.Equal(t, domain, tunnel.Domain)
+	assert.Equal(t, httpPort, tunnel.HTTPPort)
 
 	// Stop tunnel
 	err = manager.StopTunnel(ctx, domain)
@@ -87,13 +104,14 @@ func TestHTTPSTunnel(t *testing.T) {
 
 	ctx := context.Background()
 	domain := "test-https.local"
-	port := 8080
-	httpsPort := 8443
+	backendPort := 8080
+	httpPort := 8181   // Use different port to avoid conflicts
+	httpsPort := 8444  // Use different port to avoid conflicts
 
-	// Start HTTPS tunnel
-	err := manager.StartTunnel(ctx, port, domain, true, httpsPort)
+	// Start HTTPS tunnel with custom ports
+	err := manager.StartTunnelWithPorts(ctx, backendPort, domain, true, httpPort, httpsPort)
 	if err != nil {
-		t.Skipf("Skipping HTTPS test (might need privileges): %v", err)
+		t.Skipf("Skipping HTTPS test (might need certificates): %v", err)
 		return
 	}
 	defer manager.StopTunnel(ctx, domain)
@@ -113,10 +131,11 @@ func TestMultipleTunnels(t *testing.T) {
 	ctx := context.Background()
 	numTunnels := 3
 
-	// Start multiple tunnels
+	// Start multiple tunnels with different HTTP listen ports to avoid conflicts
 	for i := 0; i < numTunnels; i++ {
 		domain := fmt.Sprintf("test-%d.local", i)
-		err := manager.StartTunnel(ctx, 8080+i, domain, false, 8443+i)
+		// Use StartTunnelWithPorts to specify different HTTP and HTTPS ports
+		err := manager.StartTunnelWithPorts(ctx, 8080+i, domain, false, 8080+100+i, 8443+i)
 		require.NoError(t, err)
 	}
 
@@ -146,14 +165,14 @@ func TestErrorCases(t *testing.T) {
 		{
 			name: "Invalid port",
 			fn: func() error {
-				return manager.StartTunnel(ctx, -1, "test.local", false, 8443)
+				return manager.StartTunnelWithPorts(ctx, -1, "test.local", false, 8200, 8500)
 			},
 			wantErr: true,
 		},
 		{
 			name: "Empty domain",
 			fn: func() error {
-				return manager.StartTunnel(ctx, 8080, "", false, 8443)
+				return manager.StartTunnelWithPorts(ctx, 8080, "", false, 8201, 8501)
 			},
 			wantErr: true,
 		},
@@ -168,11 +187,11 @@ func TestErrorCases(t *testing.T) {
 			name: "Duplicate tunnel",
 			fn: func() error {
 				domain := "duplicate.local"
-				err := manager.StartTunnel(ctx, 8080, domain, false, 8443)
+				err := manager.StartTunnelWithPorts(ctx, 8080, domain, false, 8202, 8502)
 				if err != nil {
 					return err
 				}
-				return manager.StartTunnel(ctx, 8081, domain, false, 8444)
+				return manager.StartTunnelWithPorts(ctx, 8081, domain, false, 8203, 8503)
 			},
 			wantErr: true,
 		},
